Users dashboards

+++ app/Http/Controllers/DashboardController.php
@@ -6,19 +6,50 @@ use App\Models\Dashboard;
 use Illuminate\Http\Request;
 use App\Http\Requests\DashboardRequest;
 use Illuminate\Support\Facades\Auth;
+use App\Models\User;
+use App\Policies\DashboardPolicy;
+
 
 class DashboardController extends Controller
 {
+
+
     public function index($locale)
     {
-        $dashBoards = Dashboard::all();
+        $this->authorize('viewAny', Dashboard::class);
+
+        $user = Auth::user();
+        $userAssignedDashboards = $user->userDashboards()->with('owner')->get()->map(function ($d) use ($user) {
+            $d->can = [
+                'update' => $user->can('update', $d),
+                'delete' => $user->can('delete', $d),
+                'view'   => $user->can('view', $d),
+            ];
+            return $d;
+        });
+
+        $userCreatedDashboards = ($user->type === 'admin') || ($user->type === 'teacher') ? $user->dashboards()->with('owner')->get()->map(function ($d) use ($user) {
+            $d->can = [
+                'update' => $user->can('update', $d),
+                'delete' => $user->can('delete', $d),
+                'view'   => $user->can('view', $d),
+            ];
+            return $d;
+        }) : null;
 
         return inertia('Dashboard/Index', [
-            'dashboards' => $dashBoards,
+            'assignedDashboards' => $userAssignedDashboards,
+            'createdDashboards' => $userCreatedDashboards ?? [],
             'locale' => $locale,
             'translations' => [
+                'dashboard' => __('app.dashboard'),
                 'dashboards' => __('app.dashboards'),
                 'all_dashboards' => __('app.all_dashboards'),
+                'my_dashboards' => __('app.my_dashboards'),
+                'create_dashboard' => __('app.create_dashboard'),
+                'assigned_dashboards' => __('app.assigned_dashboards'),
+                'created_dashboards' => __('app.created_dashboards'),
+                'created_by' => __('app.created_by'),
                 'no_images' => __('app.no_images'),
                 'create' => __('app.create'),
                 'name' => __('app.name'),
@@ -28,13 +59,14 @@ class DashboardController extends Controller
                 'sign_in' => __('auth.sign_in'),
                 'sign_out' => __('auth.sign_out'),
                 'edit_profile' => __('app.edit_profile')
-
             ]
         ]);
     }
 
     public function create($locale)
     {
+        $this->authorize('create', Dashboard::class);
+
         return inertia('Dashboard/Create', [
             'locale' => $locale,
             'translations' => [
@@ -58,11 +90,11 @@ class DashboardController extends Controller
 
     public function store(DashboardRequest $request, $locale)
     {
-        // $dashboard = auth()->user()->dashboards()->create($request->validated()); DODAĆ PO WPROWADZENIU AUTHA
-        $validated = $request->validated();
-        $validated['created_by'] = 21;
+        $this->authorize('create', Dashboard::class);
 
-        $dashboard = Dashboard::create($validated);
+        $validated = $request->validated();
+        $validated['created_by'] = Auth::id();
+        Dashboard::create($validated);
 
         return redirect()->route('dashboard.index', ['locale' => $locale])
             ->with('success', $this->successMessage('create', 'dashboard'));
@@ -70,12 +102,17 @@ class DashboardController extends Controller
 
     public function show($locale, Dashboard $dashboard)
     {
+        $this->authorize('view', $dashboard);
+        $dashboard->load('owner');
+
         return inertia('Dashboard/Show', [
             'dashboard' => $dashboard,
             'locale' => $locale,
             'translations' => [
                 'dashboards' => __('app.dashboards'),
                 'create' => __('app.create'),
+                'create_assignment' => __('app.create_assignment'),
+                'assign_users' => __('app.assign_users'),
                 'dashboard' => __('app.dashboard'),
                 'name' => __('app.name'),
                 'description' => __('app.description'),
@@ -90,6 +127,8 @@ class DashboardController extends Controller
 
     public function edit($locale, Dashboard $dashboard)
     {
+        $this->authorize('update', $dashboard);
+
         return inertia('Dashboard/Edit', [
             'dashboard' => $dashboard,
             'locale' => $locale,
@@ -101,6 +140,7 @@ class DashboardController extends Controller
                 'create' => __('app.create'),
                 'update' => __('app.update'),
                 'dashboard' => __('app.dashboard'),
+                'create_dashboard' => __('app.create_dashboard'),
                 'name' => __('app.name'),
                 'description' => __('app.description'),
                 'edit' => __('app.edit'),
@@ -114,6 +154,8 @@ class DashboardController extends Controller
 
     public function update(DashboardRequest $request, $locale, Dashboard $dashboard)
     {
+        $this->authorize('update', $dashboard);
+
         $validated = $request->validated();
         $dashboard->update($validated);
 
@@ -123,6 +165,8 @@ class DashboardController extends Controller
 
     public function destroy($locale, Dashboard $dashboard)
     {
+        $this->authorize('delete', $dashboard);
+
         $dashboard->delete();
 
         return redirect()->route('dashboard.index', ['locale' => $locale])
